+++ lib/components/constants.dart
@@ -11,7 +11,7 @@ import 'package:flutter/material.dart';
   const double kDefaultPadding = 15.0;
   const double kDefaultRadius = 30.0;
   const double kCardDefaultHeigth = 200 + kDefaultPadding * 2;
-  const int kMobilSize = 6;
-  const int kTabletSize = 8;
-  const int kDesktopSize = 10;
+  const int kMobilSize = 5;
+  const int kTabletSize = 6;
+  const int kDesktopSize = 8;
